Add enums to Orchestration and export them from index.ts for visibility
