--- src/orchestration/constants.ts
@@ -1,12 +1,0 @@
-export enum PAGE_ID_FORMAT {
-    PATH = 'PATH',
-    HASH = 'HASH',
-    PATH_AND_HASH = 'PATH_AND_HASH'
-}
-
-export enum TELEMETRY_TYPES {
-    ERRORS = 'errors',
-    PERFORMANCE = 'performance',
-    INTERACTION = 'interaction',
-    HTTP = 'http'
-}
